+++ packages/angular/test/base/src/app/lazy/tabs-basic/tabs-basic.component.ts
@@ -1,5 +1,5 @@
 import { Component, ViewChild } from '@angular/core';
-import { IonTabBar } from '@ionic/angular';
+import { IonTabBar } from '@ionic/angular/lazy';
 
 @Component({
   selector: 'app-tabs-basic',