+++ packages/angular/test/base/src/app/lazy/version-test/bind-component-inputs/bind-component-inputs.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input, OnInit } from "@angular/core";
-import { IonicModule } from '@ionic/angular';
+import { IonicModule } from '@ionic/angular/lazy';
 
 @Component({
   selector: 'app-bind-route',