+++ packages/angular/test/base/src/app/standalone/value-accessors/datetime/datetime.component.ts
@@ -1,6 +1,6 @@
 import { Component } from "@angular/core";
 import { FormBuilder, FormsModule, ReactiveFormsModule, Validators } from "@angular/forms";
-import { IonDatetime } from "@ionic/angular/standalone";
+import { IonDatetime } from "@ionic/angular";
 import { ValueAccessorTestComponent } from "../value-accessor-test/value-accessor-test.component";
 
 @Component({