+++ packages/vue/src/controllers.ts
@@ -4,14 +4,12 @@ import {
   alertController as alertCtrl,
   actionSheetController as actionSheetCtrl,
   loadingController as loadingCtrl,
-  pickerController as pickerCtrl,
   toastController as toastCtrl,
 } from "@ionic/core/components";
 import { defineCustomElement as defineIonActionSheetCustomElement } from "@ionic/core/components/ion-action-sheet.js";
 import { defineCustomElement as defineIonAlertCustomElement } from "@ionic/core/components/ion-alert.js";
 import { defineCustomElement as defineIonLoadingCustomElement } from "@ionic/core/components/ion-loading.js";
 import { defineCustomElement as defineIonModalCustomElement } from "@ionic/core/components/ion-modal.js";
-import { defineCustomElement as defineIonPickerCustomElement } from "@ionic/core/components/ion-picker-legacy.js";
 import { defineCustomElement as defineIonPopoverCustomElement } from "@ionic/core/components/ion-popover.js";
 import { defineCustomElement as defineIonToastCustomElement } from "@ionic/core/components/ion-toast.js";
 
@@ -72,13 +70,6 @@ const loadingController = /*@__PURE__*/ createController(
   loadingCtrl
 );
 
-/**
- * @deprecated Use the inline ion-picker component instead.
- */
-const pickerController = /*@__PURE__*/ createController(
-  defineIonPickerCustomElement,
-  pickerCtrl
-);
 const toastController = /*@__PURE__*/ createController(
   defineIonToastCustomElement,
   toastCtrl
@@ -90,6 +81,5 @@ export {
   alertController,
   actionSheetController,
   loadingController,
-  pickerController,
   toastController,
 };